Replace display feedback strength setting with on screen display setting
Replacing the display feedback strength setting with a better on screen display setting with the following options-
None: No information will be displayed on screen.
MessageOnly: Only a simple message will be displayed when feedback is incoming and when they trigger.
DistanceBasedPercentage: A simple message plus a 1-100% will be displayed calculated based on the min/max distances provided to trigger the feedback.
FeedbackIntensityAndDuration: A simple message plus specific information about the feedback intensity and duration will be displayed.