Improve Modifiers Compatability
Adding handling of calls to ApplySaveState moving the player so it doesn't cause feedback to trigger incorrectly- this makes the punishment mod work well enough together with the Rising Lava and Fall Damage modifiers.

Also adding a zOrder to the punishment mod text entities so they render over stuff like the Lava.